back_django_insta/pyproject.toml
[tool.poetry]
name = "back-django-insta"
version = "0.1.0"
description = ""
authors = ["jaeupsu <hyeongs2323@gmail.com>"]
readme = "README.md"
packages = [{include = "back_django_insta"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.5"
selenium = "^4.7.2"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
